requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-static==0.0.6
Django==4.0.1
django-debug-toolbar==3.2.4
django-filter==21.1
django-graphql-jwt==0.3.4
django-import-export==2.7.1
django-js-asset==1.2.2
django-mptt==0.13.4
django-polymorphic==3.1.0
django-polymorphic-tree==2.1
django-spaghetti-and-meatballs==0.4.2
django-tag-parser==3.2
djongo==1.3.6
dnspython==2.2.0
et-xmlfile==1.1.0
graphene==2.1.9
graphene-django==2.15.0
graphene-django-optimizer==0.9.1
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.0
promise==2.3
PyJWT==2.3.0
pymongo==3.12.3
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.2.4
static3==0.7.0
tablib==3.2.0
text-unidecode==1.3
xlrd==2.0.1
xlwt==1.3.0
